requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.0.7
django-debug-toolbar==1.10.1
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Flask==1.0.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
